Entities Service

+++ src/main/java/myUniversityPack/EntityService/DatabaseService.java
@@ -5,7 +5,7 @@ import java.util.Collection;
 public interface DatabaseService<T> {
     public boolean doSave(T t);
     public boolean remove(T t);
-    default T findById(long id){return null;}
+    default T findById(int id){return null;}
     public T findById(String matricola);
     public Collection<T> findAll();
 }
